file_manipulator/pyproject.toml
[project]
name = "gpti-proyecto"
version = "0.1.0"
description = ""
authors = [
    {name = "Borja MPM",email = "bmarquezdelaplata@uc.cl"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pypdf2 (>=3.0.1,<4.0.0)",
    "pdf2image (>=1.17.0,<2.0.0)",
    "pillow (>=11.2.1,<12.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{include = "app"}]
